requirements_pdf.txt
# PDF RAG System Requirements
pypdf2
langchain
langchain-community
faiss-cpu
sentence-transformers
chromadb
tiktoken
